Move fetch_gui script to helpers and add document Edit readme and doc files with Hmi information typo on app.py
